tests/requirements.txt
django==1.8
django-nose==1.4
